Added ability to find page numbers
If you go to this PyPi URL and scroll down to the navigation buttons near the bottom, the button to the left of "next" tells us how many page numbers we need to scrape for that license.

I've added some code that can grab that, but on different pages, it is a different index. So this will need updating, but it's progress. Now we know how many pages, which is needed to automate some web scraping processes.